Move terminal notifier to it's own file
