+++ tools/notifier/__init__.py
@@ -16,9 +16,6 @@
 from __future__ import print_function, division, absolute_import
 
 from abc import ABCMeta, abstractmethod
-from os.path import basename
-
-from ..settings import PRINT_COMPILER_OUTPUT_AS_LINK
 
 
 class Notifier(object):
@@ -103,72 +100,3 @@ class Notifier(object):
         Update a UI with a key, value pair
         """
         self.notify({'type': 'var', 'key': key, 'val': value})
-
-
-class TerminalNotifier(Notifier):
-    """
-    Writes notifications to a terminal based on a silent and verbose flag.
-    """
-
-    def __init__(self, verbose=False, silent=False):
-        self.verbose = verbose
-        self.silent = silent
-        self.output = ""
-
-    def get_output(self):
-        return self.output
-
-    def notify(self, event):
-        if self.verbose:
-            msg = self.print_notify_verbose(event)
-        else:
-            msg = self.print_notify(event)
-        if msg:
-            if not self.silent:
-                print(msg)
-            self.output += msg + "\n"
-
-    def print_notify(self, event):
-        """ Default command line notification
-        """
-        if not self.verbose and event['type'] == 'tool_error':
-            return event['message']
-
-        elif event['type'] in ['info']:
-            return event['message']
-
-        elif event['type'] == 'cc':
-            event['severity'] = event['severity'].title()
-
-            if PRINT_COMPILER_OUTPUT_AS_LINK:
-                event['file'] = getcwd() + event['file'].strip('.')
-                return '[%(severity)s] %(file)s:%(line)s:%(col)s: %(message)s' % event
-            else:
-                event['file'] = basename(event['file'])
-                return '[%(severity)s] %(file)s@%(line)s,%(col)s: %(message)s' % event
-
-        elif event['type'] == 'progress':
-            if 'percent' in event:
-                return '{} [{:>5.1f}%]: {}'.format(event['action'].title(),
-                                                  event['percent'],
-                                                  basename(event['file']))
-            else:
-                return '{}: {}'.format(event['action'].title(),
-                                      basename(event['file']))
-
-    def print_notify_verbose(self, event):
-        """ Default command line notification with more verbose mode
-        """
-        if event['type'] in ['info', 'debug']:
-            return event['message']
-
-        elif event['type'] == 'cc':
-            event['severity'] = event['severity'].title()
-            event['file'] = basename(event['file'])
-            event['mcu_name'] = "None"
-            event['target_name'] = event['target_name'].upper() if event['target_name'] else "Unknown"
-            event['toolchain_name'] = event['toolchain_name'].upper() if event['toolchain_name'] else "Unknown"
-            return '[%(severity)s] %(target_name)s::%(toolchain_name)s::%(file)s@%(line)s: %(message)s' % event
-
-        elif event['type'] == 'progress':
-            return self.print_notify(event) # standard handle
